Remove Warnings caused by signed/unsigned integer comparaisons

+++ miniupnpd/miniupnpd.c
@@ -1,4 +1,4 @@
-/* $Id: miniupnpd.c,v 1.153 2012/04/27 06:48:42 nanard Exp $ */
+/* $Id: miniupnpd.c,v 1.154 2012/04/30 13:38:21 nanard Exp $ */
 /* MiniUPnP project
  * http://miniupnp.free.fr/ or http://miniupnp.tuxfamily.org/
  * (c) 2006-2012 Thomas Bernard
@@ -541,7 +541,7 @@ parselanaddr(struct lan_addr_s * lan_addr, const char * str)
 	while(*p && *p != '/' && !isspace(*p))
 		p++;
 	n = p - str;
-	if(!isdigit(str[0]) && n < sizeof(lan_addr->ifname))
+	if(!isdigit(str[0]) && n < (int)sizeof(lan_addr->ifname))
 	{
 		/* not starting with a digit : suppose it is an interface name */
 		memcpy(lan_addr->ifname, str, n);
@@ -683,7 +683,7 @@ init(int argc, char * * argv, struct runtime_vars * v)
 	}
 	else
 	{
-		for(i=0; i<num_options; i++)
+		for(i=0; i<(int)num_options; i++)
 		{
 			switch(ary_options[i].id)
 			{
@@ -1387,13 +1387,13 @@ main(int argc, char * * argv)
 		/* Remove expired port mappings, based on UPnP IGD LeaseDuration
 		 * or NAT-PMP lifetime) */
 		if(nextruletoclean_timestamp
-		  && (timeofday.tv_sec >= nextruletoclean_timestamp))
+		  && ((unsigned int)timeofday.tv_sec >= nextruletoclean_timestamp))
 		{
 			syslog(LOG_DEBUG, "cleaning expired Port Mappings");
 			get_upnp_rules_state_list(0);
 		}
 		if(nextruletoclean_timestamp
-		  && timeout.tv_sec >= (nextruletoclean_timestamp - timeofday.tv_sec))
+		  && ((unsigned int)timeout.tv_sec >= (nextruletoclean_timestamp - timeofday.tv_sec)))
 		{
 			timeout.tv_sec = nextruletoclean_timestamp - timeofday.tv_sec;
 			timeout.tv_usec = 0;
